Fixed #18004 -- Don't silently mangle badly encoded POST filename data.
Use `force_unicode(strict=True)` when decoding POST filenames, to
trigger a loud failure with a useful error.

Don't send people down the rabbit hole trying to debug "current
transaction is aborted" errors where it was the previous statement that
triggered the error, but the previous statement is not logged anywhere
because of the encoding error.